refactor: 테스트 용 Generator 구현 클래스의 위치를 도메인으로 이동
