+++ src/main/java/model/ladder/generator/RandomStatusGenerator.java
@@ -1,6 +1,7 @@
-package model.ladder;
+package model.ladder.generator;
 
 import java.util.Random;
+import model.ladder.StepStatus;
 
 public class RandomStatusGenerator implements StepStatusGenerator {
     private static final Random random = new Random();
