+++ src/main/java/model/ladder/generator/StepStatusGenerator.java
@@ -1,4 +1,6 @@
-package model.ladder;
+package model.ladder.generator;
+
+import model.ladder.StepStatus;
 
 @FunctionalInterface
 public interface StepStatusGenerator {
