+++ src/main/java/model/ladder/generator/TestStepStatusGenerator.java
@@ -1,0 +1,22 @@
+package model.ladder.generator;
+
+import model.ladder.StepStatus;
+
+public class TestStepStatusGenerator {
+
+    public static class AlwaysConnectStatusGenerator implements StepStatusGenerator {
+
+        @Override
+        public StepStatus generate() {
+            return StepStatus.CONNECTED;
+        }
+    }
+
+    public static class AlwaysDisconnectStatusGenerator implements StepStatusGenerator {
+
+        @Override
+        public StepStatus generate() {
+            return StepStatus.DISCONNECTED;
+        }
+    }
+}
